legacy/4kagent/executor/super_resolution/tools/OSEDiff/requirements.txt
torch==2.0.1
transformers==4.28.1
einops==0.7.0
Pillow==9.5.0
PyYAML==6.0
pyiqa==0.1.10
basicsr
loralib
fairscale
diffusers==0.25.0
open-clip-torch
peft==0.10.0
natsort
numpy==1.26
huggingface-hub==0.25.2
omegaconf
torchsde
pytorch_lightning
xformers==0.0.20
tqdm